Fixed bug in rendering certain inputs
This fixes a problem with rendering inputs like this one:

PRINT ARG=d1.*

The problem was that the . was treated as a word boundary when you did the lexing.
The * is thus not output in the rendered input.  I thus changed the way that the input
is lexed so that you look for a whitespace after the argument for the keyword rather than
the word boundary